+++ Assets/uNode3/Core.Editor/ItemSelector/ItemSelector.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using System.Threading;
-using System.Xml;
 using UnityEditor;
 using UnityEditor.IMGUI.Controls;
 using UnityEngine;