+++ Assets/uNode3/Core.Editor/UIElementGraph/Views/NodeViews/BaseNodeView.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Linq;
 using System.Collections.Generic;
-using System.Reflection;
 using UnityEditor;
 using UnityEngine;
 using UnityEngine.UIElements;